Authorization

+++ app/Http/Controllers/API/V1/Admin/CategoryController.php
@@ -9,6 +9,7 @@ use App\Http\Resources\CategoryResource;
 use App\Models\Category;
 use App\Services\CategoryService;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends BaseApiController
 {
@@ -17,12 +18,14 @@ class CategoryController extends BaseApiController
     }
     public function index(): JsonResponse
     {
+        Gate::authorize('viewAny', Category::class);
         $categories = $this->categoryService->getAllCategories();
         return $this->sendResponse(CategoryResource::collection($categories), 'Categories retrieved successfully.');
     }
 
     public function store(StoreCategoryRequest $request): JsonResponse
     {
+        Gate::authorize('create', Category::class);
         $data = $request->validated();
         $imageFile = $request->file('image');
         $category = $this->categoryService->createCategory($data, $imageFile);
@@ -31,12 +34,14 @@ class CategoryController extends BaseApiController
 
     public function show(Category $category): JsonResponse
     {
+        Gate::authorize('view', $category);
         $category = $this->categoryService->showCategory($category);
         return $this->sendResponse(new CategoryResource($category), 'Category retrieved successfully.');
     }
 
     public function update(UpdateCategoryRequest $request, Category $category): JsonResponse
     {
+        Gate::authorize('update', $category);
         $data = $request->validated();
         $imageFile = $request->file('image');
         $category = $this->categoryService->updateCategory($category, $data, $imageFile);
@@ -45,6 +50,7 @@ class CategoryController extends BaseApiController
 
     public function destroy(Category $category): JsonResponse
     {
+        Gate::authorize('delete', $category);
         $this->categoryService->deleteCategory($category);
         return $this->sendResponse(message: 'Category deleted successfully.');
     }
